Revert Unit::Wait removal. Commands are not currently checked to see if the Data union will stay in the same state for the entire of the command.  Eg Building a building will use the movement Data union, and then the built section to produce a building.
If wait is to be placed in this area, it needs to be done very carefully as to
not stand on any toes.  Which it is now doing.  It is debatable that is a good
idea to put it here given the potential for it to cause the problems it is now.
I may just need clear warnings