Fix bug where Unit could stay in the cache on it's death. If a unit has a death animation, but not a corpse it would get put into the cache as it has an animation, but never removed, just released in action_die.cpp
We now check for that condition in action_die.cpp.  Comments were added to unit.cpp
to make it clear what the remove/add to cache is about.

This bug could trigger crashes at any point when a missile is fired into an area
where a unit with the animation but no Corpse died.