patch #7097 "Stop mouse scrolling in patch editor"
Stop scrolling when the mouse leaves the area that caused the scrolling. This bug was only in the patch editor; not in the map editor nor in the game. While the patch editor kept trying to scroll in some direction, it was not possible to scroll to the opposite direction using the keyboard. 

Patch by Kalle Olavi Niemitalo.