Discard NewOrder if it refers to destroyed unit
If a building has a NewOrder whose Goal points to a destroyed unit,
then clear that order when a unit is trained in that building.
Previously, Bos Wars did this only if the unit was able to execute
the order; not if it changed the order to UnitActionMove.

Test case: Tell a training camp to make the trained units harvest a
tree.  Harvest the tree or blow it up.  Wait until the corpse of the
tree disappears.  Train an assault unit in the camp.  Without this
patch, Assert(Refs && !Destroyed) fails in CUnit::RefsIncrease.

Fixes bug #29303 "Assert failed while playing 4on4.map" http://savannah.nongnu.org/bugs/?29303

Patch by Kalle Olavi Niemitalo.