Set ThisPlayer when entering the EditorEditUnit state
If you opened bottleneck.map in the map editor, clicked the
unit-editing button, selected the vault from the list, and tried
to place that over the existing vault, then the editor correctly
refused to put it there but incorrectly showed the region blocked by
the existing vault as green rather than red.  This happened because
ThisPlayer pointed to Players[0], whose type is PlayerNobody in that
map, and UnitCountSeen had therefore not checked whether the vault
was visible to that player.

If you explicitly selected player 4 in the map editor, even though
the editor had already selected it by default, then the editor did
correctly show the blocked region as red.  This happened because
when EditorCallbackButtonDown sets Editor.SelectedPlayer, it sets
ThisPlayer as well.

It seems ThisPlayer should be Players + Editor.SelectedPlayer whenever
Editor.State is EditorEditUnit.  Make it so when entering that state.