Add undo groups to the map editor
In the map editor, if the user first pressed Ctrl+Shift+t to tile the
selected patch type over the whole map but then wanted to undo that,
he had to press Ctrl+z for quite a while: first to remove all the
added patches one by one, and then to restore the original patches one
by one.  Change this so that one Ctrl+z undoes the whole operation,
and one Ctrl+y redoes it.

Likewise, if the user adds a single patch on top of others and thus
causes the editor to remove them, one Ctrl+z will undo the addition
and all the removals.

It seems there aren't any editor commands that affect multiple units.