Don't implicitly add NULLs to AnimationMap, UnitTypeMap
AnimationsByIdent returned AnimationMap[ident]; UnitTypeByIdent
likewise UnitTypeMap[ident].  std::map::operator[] always returns a
reference to a value in the map; if it doesn't find the key in the
map, it adds it there with a default value.  This could cause spurious
NULL values to be added to these maps, for example if buttons were
defined for nonexistent unit types.