Fixed crash when exiting because Graphics and GraphicHash are freed before CPatchManager
The earlier fix in bos SVN trunk r9918 on 2010-07-02 does not work
reliably with GCC 4.4.5; it seems std::map::~map does not bother
marking the map as empty.  Instead, rely on C++ constructing objects
in the order of definition and destructing them in the opposite order.

In DEBUG builds, also add assertions to catch any attempt to access
Graphics or GraphicsHash before they are constructed or after they are
destructed.