buildings and units multi-select bug fixed
Fixed a bug where if you selected multiple units around a building that can be selected with rectangle, it causes both the units and building to be selected together. This patch makes it so if you multi-select the units around the building, the building is not selected.

Patched with Code from Wyrmgus - thanks go to Andrettin.