add ptrList vector to unit manager
This changes the unit manager so that when the Alloc() function returns a new unit pointer,
it immediately puts the new CUnit* into a vector of pointers to be tracked for deallocation.
During CUnitManager::Init() which runs at the end of every game, the list of released units
is iterated through and each CUnit* is given to delete. The problem is, this doesn't seem to
keep track of all the allocations very well, which is why the game was crashing. I don't really
understand why it happened when mines were depleted, and even then sometimes it wouldn't, but
in any case it's probably a better idea to just put each allocated unit into an array right after
doing new CUnit* to make sure they all get deleted properly. I got a crash once which I can't reproduce
but it's a lot better than crashing at the end of most games.